refactor: fix file name config
